+++ ext/lib/api-v2/topics/index.js
@@ -26,7 +26,7 @@ const EDITABLE_KEYS = [
 ]
 
 class CantUploadProposal extends Error {
-  constructor () {
+  constructor() {
     super('User can not upload proposal.')
     this.status = 403
     this.code = 'UPLOAD_TOPIC_FORBIDDEN'
@@ -102,8 +102,8 @@ const automata = {
 
 
 const sendToAuthor = (req, res, next) => {
-  if (automata[req.topic.attrs.state]){
-    if (automata[req.topic.attrs.state].includes(req.body['attrs.state'])){
+  if (automata[req.topic.attrs.state]) {
+    if (automata[req.topic.attrs.state].includes(req.body['attrs.state'])) {
       switch (req.body['attrs.state']) {
         case 'pendiente':
         case 'factible':
@@ -143,33 +143,43 @@ const sendToAuthor = (req, res, next) => {
 
 
 const sendToUsers = (req, res, next) => {
-  if (automata[req.topic.attrs.state]){
-    if (automata[req.topic.attrs.state].includes(req.body['attrs.state'])){
+  if (automata[req.topic.attrs.state]) {
+    if (automata[req.topic.attrs.state].includes(req.body['attrs.state'])) {
       switch (req.body['attrs.state']) {
         case 'pendiente':
         case 'factible':
         case 'no-factible':
         case 'integrado':
-          notifier.now('subscriber-update-proposal', {
-            topic: {
-              id: req.topic['_id'],
-              mediaTitle: req.body.mediaTitle,
-              subscribers: req.topic.attrs.subscribers
-            }
-          }).then(() => {
-            next()
-          }).catch(next)
+          if (req.topic.attrs.subscribers && req.topic.attrs.subscribers.length > 0) {
+            let arrPromises = req.topic.attrs.subscribers.map(user => {
+              return notifier.now('subscriber-update-proposal', {
+                topic: {
+                  id: req.topic['_id'],
+                  mediaTitle: req.body.mediaTitle,
+                  subscriber: user
+                }
+              })
+            })
+            Promise.all(arrPromises).then(() => {
+              next()
+            }).catch(next)
+          }
           break;
         default:
-          notifier.now('subscriber-update-project', {
-            topic: {
-              id: req.topic['_id'],
-              mediaTitle: req.body.mediaTitle,
-              subscribers: req.topic.attrs.subscribers
-            }
-          }).then(() => {
-            next()
-          }).catch(next)
+          if (req.topic.attrs.subscribers && req.topic.attrs.subscribers.length > 0) {
+            let arrPromises = req.topic.attrs.subscribers.map(user => {
+              return notifier.now('subscriber-update-project', {
+                topic: {
+                  id: req.topic['_id'],
+                  mediaTitle: req.body.mediaTitle,
+                  subscriber: user
+                }
+              })
+            })
+            Promise.all(arrPromises).then(() => {
+              next()
+            }).catch(next)
+          }
           break;
       }
     } else {
@@ -183,9 +193,9 @@ const sendToUsers = (req, res, next) => {
 const subscribeUser = (req, res, next) => {
   let arrayUpdated = req.topic.attrs.subscribers
   let suscribed = false
-  if(arrayUpdated){
-    if(arrayUpdated.includes(req.user.id)){
-      arrayUpdated = arrayUpdated.filter( s => {
+  if (arrayUpdated) {
+    if (arrayUpdated.includes(req.user.id)) {
+      arrayUpdated = arrayUpdated.filter(s => {
         return s != req.user.id
       })
     } else {
@@ -217,39 +227,39 @@ const subscribeUser = (req, res, next) => {
 const goToNextRoute = (req, res, next) => next('route')
 
 app.post('/topics',
-middlewares.users.restrict,
-middlewares.forums.findFromBody,
-middlewares.forums.privileges.canCreateTopics,
-purgeBody,
-sendToAdmin,
-goToNextRoute)
+  middlewares.users.restrict,
+  middlewares.forums.findFromBody,
+  middlewares.forums.privileges.canCreateTopics,
+  purgeBody,
+  sendToAdmin,
+  goToNextRoute)
 
 app.put('/topics/:id',
-middlewares.users.restrict,
-middlewares.topics.findById,
-middlewares.forums.findFromTopic,
-middlewares.forums.privileges.canCreateTopics,
-middlewares.topics.privileges.canEdit,
-purgeBody,
-sendToAuthor,
-sendToUsers,
-goToNextRoute)
+  middlewares.users.restrict,
+  middlewares.topics.findById,
+  middlewares.forums.findFromTopic,
+  middlewares.forums.privileges.canCreateTopics,
+  middlewares.topics.privileges.canEdit,
+  purgeBody,
+  sendToAuthor,
+  sendToUsers,
+  goToNextRoute)
 
 app.post('/topics/:id/subscribe',
-// middlewares.users.restrict,
-middlewares.topics.findById,
-middlewares.forums.findFromTopic,
-subscribeUser)
+  // middlewares.users.restrict,
+  middlewares.topics.findById,
+  middlewares.forums.findFromTopic,
+  subscribeUser)
 
 app.get('/all-tags',
-(req, res, next) => {
-  try{
-    api.tag.all(function (err, tags) {
-    if (err) return _handleError(err, req, res)
-    res.status(200).json(tags.map(expose('id name')))
-    })
-  } catch(err){
-    next(err)
-  }
-})
+  (req, res, next) => {
+    try {
+      api.tag.all(function (err, tags) {
+        if (err) return _handleError(err, req, res)
+        res.status(200).json(tags.map(expose('id name')))
+      })
+    } catch (err) {
+      next(err)
+    }
+  })
