+++ ext/lib/notifier/jobs/update-project/template.js
@@ -14,7 +14,7 @@ module.exports = ({
   <p>${topic.authorName},</p>
   <p>¡El estado de tu proyecto "${topic.mediaTitle}" ha sido actualizado!
   <br>Podrás ver los cambios haciendo click acá:
-  <br><a href="https://presupuestoparticipativo.vicentelopez.gob.ar/propuesta/topic/${topic.id}">https://presupuestoparticipativo.vicentelopez.gob.ar/propuesta/topic/${topic.id}</a></p>
+  <br><a href="https://presupuestoparticipativo.vicentelopez.gob.ar/proyecto/topic/${topic.id}">https://presupuestoparticipativo.vicentelopez.gob.ar/proyecto/topic/${topic.id}</a></p>
   <p>Muchas gracias por tu aporte. ¡Sigamos mejorando juntos cada barrio de Vicente López!</p>
   <p>
     PRESUPUESTO PARTICIPATIVO DE VICENTE LOPEZ<br>
